Provide defaults
There is no warning when doing something like
```php
        $this->configManager->changeConfigs(
            [
                [
                'value' => $value,
                'path' => 'some/config/path',
                ]
            ]
        );
```
The above config is not set, as it's not valid: it's missing 2 keys - `scope_type` and `scope_config`.

This has bitten me two times now. 
Either this (setting defaults) or open the doors of hell so that the developer knows something's wrong (an exception maybe).